gitlab-deployer/requirements-desktop.txt
# Desktop Build Dependencies for BuildForever Win32 Executable
# These are additional dependencies required for building the standalone Windows app

# PyInstaller - Package Python applications into standalone executables
pyinstaller>=6.0.0

# pywebview - Create native desktop windows with HTML/CSS/JS
# This provides a native Windows window instead of requiring a browser
pywebview>=4.4.0

# Note: On Windows, pywebview automatically installs:
# - pythonnet (for .NET integration with Windows webview)
# - pywin32 (for Windows API access)

# These dependencies are supplementary to the base requirements.txt
# and are specifically needed to build the standalone Windows executable.
